pyproject.toml
[project]
name = "text2text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=3.6.0",
    "ipython>=8.37.0",
    "iso639-lang>=2.6.3",
    "sacrebleu>=2.5.1",
    "seaborn>=0.13.2",
    "unbabel-comet>=v2.0.2",
    "vllm>=0.9.1",
]
[project.scripts]
prompt = "text2text.prompt:main"
[tool.uv]
package = true
